# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "卡密授权系统 - 客户端 API",
    "description": "终端软件集成用的卡密验证 API。字段名采用大写开头（Softid/Card/Mac/Token）。所有接口均为 POST，返回 JSON；除 /announcement 返回 text/plain 外，业务失败以 errcode 字段返回错误码（含义见 /api/public/error-codes 或 README 错误码表）。",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/login": {
      "post": {
        "summary": "卡密登录",
        "description": "卡密 + 机器码绑定登录。首次激活计算有效期并绑定机器码；同设备复登轮换新会话（旧 token 立即失效）；异设备登录返回 -1011。",
        "operationId": "cardLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "Softid",
                  "Card",
                  "Mac"
                ],
                "properties": {
                  "Softid": {
                    "$ref": "#/components/schemas/Softid"
                  },
                  "Card": {
                    "type": "string",
                    "description": "卡密",
                    "example": "Ab3xK9mPq2Rt5W"
                  },
                  "Mac": {
                    "type": "string",
                    "description": "机器码（任意设备唯一标识，大小写不敏感）",
                    "example": "AA-BB-CC-DD-EE-FF"
                  },
                  "Version": {
                    "type": "string",
                    "description": "客户端版本号（应用开启强制更新时校验）",
                    "example": "1.0.0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功返回 16 位会话 token；失败返回 errcode（-1001/-1004/-1005/-1006/-1007/-1008/-1010/-1011/-1012）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "maxLength": 16,
                      "example": "SNTmfDNVC-Aa2F0l"
                    },
                    "errcode": {
                      "$ref": "#/components/schemas/Errcode"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "卡密登出",
        "description": "释放会话（token 置空），机器码绑定保留。",
        "operationId": "cardLogout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功 result=1；失败返回 errcode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/heartbeat": {
      "post": {
        "summary": "心跳保活",
        "description": "校验有效会话并将 token 有效期延长 24 小时。返回 -1002 表示会话已失效，客户端应重新走 /login；返回 -1005 表示卡密已到期。",
        "operationId": "heartbeat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功 result=1；失败返回 errcode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/expiry": {
      "post": {
        "summary": "获取到期时间（需登录）",
        "operationId": "getExpiry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功返回 expires_at（本地时间 YYYY-MM-DD HH:mm:ss）；失败返回 errcode",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expires_at": {
                      "type": "string",
                      "example": "2026-09-20 13:58:20"
                    },
                    "errcode": {
                      "$ref": "#/components/schemas/Errcode"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Softid": {
        "type": "string",
        "pattern": "^[A-Za-z0-9]{18}$",
        "description": "18 位软件标识（后台创建应用时生成）",
        "example": "1460MREWAFMB3XQFEP"
      },
...
      "TokenRequest": {
        "type": "object",
        "required": [
          "Softid",
          "Card",
          "Token"
        ],
        "properties": {
          "Softid": {
            "$ref": "#/components/schemas/Softid"
          },
          "Card": {
            "type": "string",
            "description": "卡密"
          },
          "Token": {
            "type": "string",
            "description": "登录成功返回的会话 token"
          }
        }
      },
      "ResultResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string",
            "enum": [
              "1"
            ],
            "description": "成功固定为 1"
          },
          "errcode": {
            "$ref": "#/components/schemas/Errcode"
          }
        }
      },
      "Errcode": {
        "type": "string",
        "description": "错误码，含义：-1001 参数错误 / -1002 认证失败 / -1004 卡密不存在 / -1005 卡密已过期 / -1006 卡密已禁用 / -1007 应用不存在 / -1008 需强制更新 / -1009 服务器内部错误 / -1010 机器码不匹配 / -1011 已在其他设备登录 / -1012 管理员激活配额已满"
      }
    }
  }
}